chore(connector): import types from accounts service
