+++ .github/workflows/lint_test_build.yml
@@ -85,6 +85,7 @@ jobs:
       - run: yarn workspace connector lint
       - run: yarn workspace connector format:check
       - run: yarn install --immutable --immutable-cache
+      - run: yarn build
       - run: yarn workspace connector test
 
   frontend:
@@ -103,6 +104,7 @@ jobs:
   build:
     runs-on: ubuntu-latest
     needs:
+      - accounts
       - backend
       - connector
       - frontend
