+++ packages/accounts/package.json
@@ -1,5 +1,6 @@
 {
   "name": "accounts",
+  "main": "dist/index.js",
   "scripts": {
     "lint": "yarn lint:local",
     "format": "yarn format:local",
