+++ packages/accounts/src/index.ts
@@ -154,5 +154,6 @@ if (!module.parent) {
 }
 
 export * from './app'
+export * from './config'
 export * from './errors'
-export * from './services'
+export * from './types'
