+++ packages/accounts/src/types/ilpBalance.ts
@@ -1,0 +1,16 @@
+// interface IlpBalanceChildren {
+//   availableCredit: bigint
+//   totalLent: bigint
+// }
+
+// interface IlpBalanceParent {
+//   availableCreditLine: bigint
+//   totalBorrowed: bigint
+// }
+
+export interface IlpBalance {
+  id: string
+  balance: bigint
+  // children?: IlpBalanceChildren
+  // parent?: IlpBalanceParent
+}
