+++ packages/connector/src/services/core/controllers/client.ts
@@ -3,7 +3,7 @@ import { RafikiContext, RafikiMiddleware } from '../rafiki'
 import { modifySerializedIlpPrepare } from '../lib'
 //import { AxiosClient } from '../services/client/axios'
 import { sendToPeer as sendToPeerDefault } from '../services/client'
-import { IlpAccount } from '../services'
+import { IlpAccount } from 'accounts'
 
 export interface ClientControllerOptions {
   sendToPeer?: (
