+++ packages/connector/src/services/core/middleware/account.ts
@@ -2,7 +2,7 @@ import { Errors } from 'ilp-packet'
 import { RafikiContext, RafikiMiddleware } from '../rafiki'
 import { AuthState } from './auth'
 import { AccountNotFoundError } from '../errors'
-import { IlpAccount } from '../services'
+import { IlpAccount } from 'accounts'
 
 export function createAccountMiddleware(): RafikiMiddleware {
   return async function account(
