+++ packages/connector/src/services/core/middleware/auth.ts
@@ -1,4 +1,4 @@
-import { IlpAccount } from '../services'
+import { IlpAccount } from 'accounts'
 
 export interface AuthState {
   account?: IlpAccount
