+++ packages/connector/src/services/core/services/index.ts
@@ -1,4 +1,3 @@
-export * from './accounts'
 export * from './client'
 export * from './logger'
 //export * from './router'
