+++ packages/connector/src/services/core/test/middleware/token-auth-middleware.test.ts
@@ -2,6 +2,7 @@
 import Faker from 'faker'
 import { createContext } from '../../utils'
 import { createTokenAuthMiddleware } from '../../middleware'
+import { MockAccountsService } from '../mocks/accounts-service'
 import { RafikiContext } from '../../rafiki'
 import { PeerAccountFactory, RafikiServicesFactory } from '../../factories'
 
@@ -37,6 +38,7 @@ describe('Token Auth Middleware', function () {
     })
 
     test('succeeds for valid token and binds data to context', async () => {
+      const accounts = new MockAccountsService('test.rafiki')
       const ctx = createContext<unknown, RafikiContext>({
         req: {
           headers: {
@@ -44,7 +46,7 @@ describe('Token Auth Middleware', function () {
             authorization: 'Bearer asd123'
           }
         },
-        services: RafikiServicesFactory.build({})
+        services: RafikiServicesFactory.build({ accounts })
       })
       const account = PeerAccountFactory.build({
         accountId: 'alice',
@@ -58,7 +60,7 @@ describe('Token Auth Middleware', function () {
           }
         }
       })
-      await ctx.services.accounts.createAccount(account)
+      await accounts.createAccount(account)
 
       await createTokenAuthMiddleware()(ctx, async () => {})
       expect(ctx.state.account).toBe(account)
