--- packages/connector/src/services/core/types/index.ts
@@ -1,1 +1,0 @@
-//export * from './account'
